Lesson 5 accept tests

+++ app/controllers/questions_controller.rb
@@ -1,14 +1,18 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: [:index, :show]
+
   def index
     @questions = Question.all
   end
 
   def show
-
+    @question = Question.find(params[:id])
+    @answer = @question.answers.new
   end
 
   def new
-
+    @question = Question.new
+    @answer = @question.answers.new
   end
 
   def edit
@@ -16,10 +20,10 @@ class QuestionsController < ApplicationController
   end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.new(question_params)
 
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your question succesfully created'
     else
       render :new
     end
@@ -34,8 +38,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author_of?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Question deleted'
+    else
+      redirect_to question_path(question), notice: "You can't dalete this question"
+    end
   end
 
   private
